fix(speech): bring native.rs back under the module size and clippy gates
This branch grew native.rs from 797 to 851 lines, past the repository's 800
line budget, and left run_sherpa_worker_after_load taking eight positional
arguments where clippy allows seven. Neither shows up locally without running
the CI checklist by hand, because the fork has GitHub Actions disabled.

The worker thread moves to its own module and its shared state travels as a
WorkerContext instead of five loose parameters. With the lib compiling again,
clippy reached the test target and flagged an assertion over two constants
that folds to assert!(true); it becomes a compile-time assertion next to the
constant it guards, which is what it was trying to say.
